requirements.txt
certifi==2018.11.29
Django==2.1.5
djangorestframework==3.9.0
pytz==2018.7
